+++ misc/table_driven_lsc.h
@@ -1,0 +1,34 @@
+/* Copyright (C) 2009-2021 Greenbone Networks GmbH
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+/**
+ * @file table_drive_lsc.h
+ * @brief Header file for module table_driven_lsc.
+ */
+
+#ifndef TABLE_DRIVEN_LSC_H
+#define TABLE_DRIVEN_LSC_H
+
+#include <glib.h>
+
+gchar *
+make_table_driven_lsc_info_json_str (const char *, const char *, const char *,
+                                     const char *, const char *);
+
+#endif // TABLE_DRIVEN_LSC_H